Youtube_to_Blog_Generator/requirements.txt
crewai
crewai_tools
python-dotenv
langchain-huggingface
youtube-transcript-api
yt_dlp
google-api-python-client
streamlit